# src/main/resources/static/openapi/StudentService.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Student Service",
    "description": "Student Service will provide all API's required for student user management",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9080",
      "description": "Local"
    }
  ],
  "paths": {
    "/student": {
      "post": {
        "tags": [
          "Student"
        ],
        "description": "Create a Student",
        "operationId": "createStudent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/student/{id}": {
      "get": {
        "tags": [
          "Student"
        ],
        "description": "Get Student by Id",
        "operationId": "getStudents",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/student/grade/{grade}": {
      "get": {
        "tags": [
          "Student"
        ],
        "description": "Get Student by Grade",
        "operationId": "getStudents_1",
        "parameters": [
          {
            "name": "grade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Student": {
        "required": [
          "grade",
          "mobile",
          "name",
          "school"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Student Id",
            "format": "int64"
          },
          "name": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "description": "Student Name",
            "example": "Nathasha Namjith"
          },
          "grade": {
            "type": "string",
            "description": "Student Grade"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile Number",
            "example": "+971 509834759"
          },
          "school": {
            "type": "string",
            "description": "School Name",
            "example": "Gems Legacy"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "msg": {
            "type": "string"
          },
          "field": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "path": {
            "type": "string"
          }
        }
      }
    }
  }
}